refactor: ajout du dossier sections et deplacement du composant hero et about dans ce dossier + deplacement du code concernant mon profil dans un composant profil
