--- index.html
@@ -1,85 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <link rel="preconnect" href="https://fonts.googleapis.com">
-  <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-  <link href="https://fonts.googleapis.com/css2?family=Gruppo&family=Montserrat:wght@100;400;500&family=Open+Sans:wght@400;600;700&family=Staatliches&display=swap" rel="stylesheet">
-  <link rel="stylesheet" href="assets/styles/index.css">
-  <title>Mon portfolio</title>
-</head>
-
-<body>
-  <header>
-    <a href="#home"><img class="logo" src="./assets/img/logo.jpeg" alt="logo des initiales KS"></a>
-    <nav>
-      <ul>
-        <li><a href="#home">Accueil</a></li>
-        <li><a href="#about">À Propos</a></li>
-        <li><a href="#projects">Projets</a></li>
-        <li><a href="#skills">Compétences</a></li>
-        <li><a href="#contact">Contact</a></li>
-      </ul>
-    </nav>
-  </header>
-  <main>
-    <section id="home">
-      <div>
-        <h1>Bonjour, <br><span class="name-text">Je suis </span></h1>
-      </div>
-    </section>
-    <section id="about">
-      <div class="photo">
-        <img src="./assets/img/Nina Carducci-560-min.png" alt="photo de katleen">
-      </div>
-      <div class="about-me">
-        <h2>À propos de moi</h2>
-        <p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Nobis rem velit alias labore, cupiditate, doloremque excepturi ipsa laudantium debitis rerum autem aspernatur nisi itaque accusantium dolor aliquid, optio fuga facere.</p>
-      </div>
-    </section>
-    <section id="projects"></section>
-    <section id="skills">
-      <h2>Compétences</h2>
-      <div class="tabs">
-        <div class="tab-registers">
-          <button class="active-tab">Tous</button>
-          <button>Langages</button>
-          <button>Framworks</button>
-          <button>Outils</button>
-          <button>BDD</button>
-        </div>
-        <div class="tab-bodies">
-          <div class="hard-skills">
-            <ul>
-              <li><img src="./assets/img/html-logo.png" alt="logo html"></li>
-              <li><img src="./assets/img/css-logo.png" alt="logo css"></li>
-              <li><img src="./assets/img/js-logo.png" alt="logo javaScript"></li>
-              <li><img src="./assets/img/react-logo.png" alt="logo react"></li>
-              <li><img src="./assets/img/node-logo.png" alt="logo node"></li>
-              <li><img src="./assets/img/express-logo.png" alt="logo expressJS"></li>
-              <li><img src="./assets/img/mongoDB-logo.png" alt="logo mongoBD"></li>
-              <li><img src="./assets/img/figma-logo.png" alt="logo figma"></li>
-              <li><img src="./assets/img/git-logo.png" alt="logo git"></li>
-              <li><img src="./assets/img/github-logo.png" alt="logo github"></li>
-            </ul>
-          </div>
-        </div>
-    </section>
-    <section id="contact">
-      <h2>Contact</h2>
-      <form action="">
-        <label for="name">Nom</label>
-        <input type="text" id="name">
-        <label for="firstname">Prénom</label>
-        <input type="text" id="firstname">
-        <label for="email">Email</label>
-        <input type="email" id="email">
-        <label for="message">Message</label>
-        <textarea name="message" id="message" cols="30" rows="10"></textarea>
-        <button type="submit">Envoyer</button>
-      </form>
-    </section>
-  </main>
-</body>
-</html>
