+++ src/components/Profil.js
@@ -1,0 +1,23 @@
+import React from "react";
+
+export const Profil = () => {
+  return (
+    <>
+      <div>
+        <img
+          src="/assets/img/profil.png"
+          alt="Photo de profil de Katleen SORHAINDO"
+        />
+      </div>
+      <h2>À propos de moi</h2>
+      <p className="py-4">
+        Lorem ipsum dolor sit amet consectetur, adipisicing elit. Fugiat iure
+        possimus qui illo commodi eligendi repellat itaque. Assumenda ipsa
+        nostrum quibusdam magnam, quidem vel consectetur quos dolores
+        consequuntur cumque inventore.
+      </p>
+    </>
+  );
+};
+
+export default Profil;
